pyproject.toml
[project]
name = "bookdit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "atproto>=0.0.61",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-fasthtml>=0.12.24",
]
